remove cpu preprocess code
